invoice-app/server/requirements.txt
PyMuPDF==1.22.5
pytesseract==0.3.10
Pillow==9.5.0
requests==2.31.0
pdfminer.six==20250624